#12 updater.getUpdatesInFolder() returns errors.
updater.getUpdatesInFolder() now returns error, as requested in the review.
Not sure yet, what is the best way to handle occurring errors. Path errors will occur when there is no update in the checked folder. This can be a valid scenario in a use case tho, and should not be treated as a malfunction. Furthermore, if there is no update in the current major Folder, there can still be one in the latest Major Folder and vice versa. Therefore decided to just log the error for now.